+++ web/src/index.css
@@ -113,58 +113,11 @@ blockquote {
     font-style: italic;
 }
 
-.App-header {
-    background-color: #24344f;
-    position: fixed;
-    top: 0;
-    width: 100%;
-    z-index: 999;
-    padding-top: 10px;
-    padding-bottom: 10px;
-    /* height: 50px; */
-}
-
-.metamist {
-    font-family: 'Raleway', sans-serif;
-    color: white;
-    font-size: 28px;
-    /* display: inline-block; */
-}
-
-.header {
-    background-color: #24344f;
-    display: flex;
-    align-items: center;
-    padding: 0px 80px 0px 80px;
-}
-
-.metamist-img {
-    vertical-align: bottom;
-}
-
-.navbarLink {
-    color: #f2f2f2;
-    padding: 0 16px;
-    text-decoration: none;
-    font-size: 17px;
-    display: inline-block;
-}
-
-.navbarIcon {
-    padding: 0 8px;
-    display: inline-block;
-}
-
 .body {
     /* padding: 10px 20px;
   max-width: 768px;
   margin: calc(50px + 1vmin) auto 0 auto; */
     padding: 80px;
-}
-
-.App-header a {
-    color: white;
-    text-decoration: none;
 }
 
 @media screen and (min-width: 1200px) {
