+++ web/src/shared/components/Header/NavBar.css
@@ -1,0 +1,46 @@
+.App-header {
+    background-color: #24344f;
+    position: fixed;
+    top: 0;
+    width: 100%;
+    z-index: 999;
+    padding-top: 10px;
+    padding-bottom: 10px;
+    /* height: 50px; */
+}
+
+.metamist {
+    font-family: 'Raleway', sans-serif;
+    color: white;
+    font-size: 28px;
+    /* display: inline-block; */
+}
+
+.header {
+    background-color: #24344f;
+    display: flex;
+    align-items: center;
+    padding: 0px 80px 0px 80px;
+}
+
+.metamist-img {
+    vertical-align: bottom;
+}
+
+.navbarLink {
+    color: #f2f2f2;
+    padding: 0 16px;
+    text-decoration: none;
+    font-size: 17px;
+    display: inline-block;
+}
+
+.navbarIcon {
+    padding: 0 8px;
+    display: inline-block;
+}
+
+.App-header a {
+    color: white;
+    text-decoration: none;
+}
